// src/clients/dz-dialect-training-api/model/get-exercise-exam-response-dto.ts
/**
 * DzDialect Training API
 * DzDialect Training API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetExerciseExamQuestionResponseDto } from './get-exercise-exam-question-response-dto';
import { GetExerciseExamCurrentResponseDto } from './get-exercise-exam-current-response-dto';
import { GetExerciseExamResultResponseDto } from './get-exercise-exam-result-response-dto';


export interface GetExerciseExamResponseDto { 
    id: string;
    name: string;
    courseId: string;
    description: string;
    questions: Array<GetExerciseExamQuestionResponseDto>;
    current?: GetExerciseExamCurrentResponseDto;
    result?: GetExerciseExamResultResponseDto;
    order: number;
    createdAt: string;
    updatedAt: string;
}